refactor(process-options): relocate process options from utils to command executor space
